logging / fixes

+++ pyacq/core/processspawner.py
@@ -1,6 +1,7 @@
-
 import sys
 import subprocess
+import atexit
+import logging
 
 from .rpc import RPCClient
 
@@ -28,17 +29,27 @@ class ProcessSpawner:
         bootstrap = bootstrap_template.format(class_name=class_name, args=args)
         executable = sys.executable
         self.proc = subprocess.Popen((executable, '-c', bootstrap))
+        logging.info("Spawn process: %d", self.proc.pid)
         self.client = RPCClient(name, addr)
         assert self.client.ping() == 'pong', "Failed to start process."
+        
+        # automatically shut down process when we exit. 
+        atexit.register(self.kill)
 
     def wait(self):
         self.proc.wait()
 
     def kill(self):
+        if self.proc.poll() is not None:
+            return
+        logging.info("Kill process: %d", self.proc.pid)
         self.proc.kill()
         self.proc.wait()
 
     def stop(self):
+        if self.proc.poll() is not None:
+            return
+        logging.info("Close process: %d", self.proc.pid)
         self.client.close()
         self.proc.wait()
